created API authentication

+++ backend/controllers/goalController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler');
 const Goal = require('../database/models/goalModel');
 
 const getGoals = asyncHandler(async(req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({user: req.user.id});
   res.status(200).json(goals);
 })
 
@@ -13,7 +13,8 @@ const setGoal = asyncHandler(async(req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
   })
   res.status(200).json(goal);
 })
@@ -24,6 +25,18 @@ const updateGoal = asyncHandler(async(req, res) => {
   if(!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = req.user;
+
+  if(!user) {
+    res.status(401);
+    throw new Error('User not found.');
+  }
+
+  if(goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized.')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -37,6 +50,18 @@ const deleteGoal = asyncHandler(async(req, res) => {
   if(!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = req.user;
+
+  if(!user) {
+    res.status(401);
+    throw new Error('User not found.');
+  }
+
+  if(goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized.')
   }
 
   await Goal.deleteOne(goal);
